Merge in EADPGHS/light-4j from master to sync

* commit '5651388b8bd031612a7ea02ec05b5b46bd7b994b': (70 commits)
  update version.jackson to 2.19.4
  fixes #2678 update RuleLoaderStartupHook to switch from market to service
  fixes #2677 regenerate ruleloader-config schema and events
  fixes #2676 Add a constant jsonrpc_id for audit info
  fixes #2675 add token exchang type ccac and msal to Constants
  fixes #2674 Add a new test case to convert embeded json for JsonMapper
  fixes #2672 upgrade junit to version 5
  upgrade to version 2.3.4-SNAPSHOT after release in master branch
  light-bot checkin CHANGELOG.md
  upgrade to version 2.3.3 before release in master branch
  fixes #2671 Update config YamlGenerator to add injection flag
  Additional Refactor For Token Exchange Handler (#2668)
  fixes #2666 remove unused org.apache package from config module
  fixes #2665 fix the flaky test in metrics-config
  fixes #2664 rename SimpleConnectHolder to SimpleConnectionState
  fixes #1950 Unit test in client module fails on some systems because it runs too fast
  fixes #2023 Remove unused code in SimpleUndertowConnectionMaker
  fixes #2028 Ensure leaked connections are closed
  fixes #2053 Ensure XnioSSL and XnioWorker resources are created only once
  fixes #2375 fix the ConsulRegistry to support optional tag for lookup
  ...

+++ server/src/test/java/com/networknt/server/TestMergeStatusConfig.java
@@ -17,9 +17,7 @@ package com.networknt.server;
 
 import com.networknt.config.Config;
 import com.networknt.status.Status;
-import junit.framework.TestCase;
-import org.junit.*;
-import org.junit.runners.MethodSorters;
+import org.junit.jupiter.api.*;
 
 import java.io.File;
 import java.io.IOException;
@@ -27,17 +25,18 @@ import java.lang.reflect.Field;
 import java.util.HashMap;
 import java.util.Map;
 
-@FixMethodOrder(MethodSorters.NAME_ASCENDING)
-public class TestMergeStatusConfig extends TestCase {
+import static org.junit.jupiter.api.Assertions.fail;
+
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+@TestMethodOrder(MethodOrderer.MethodName.class)
+public class TestMergeStatusConfig {
 
     private Config config = null;
 
     final String homeDir = System.getProperty("user.home");
 
-    @BeforeClass
+    @BeforeAll
     public void setUp() throws Exception {
-        super.setUp();
-
         config = Config.getInstance();
         setExternalizedConfigDir(homeDir);
 
@@ -50,64 +49,73 @@ public class TestMergeStatusConfig extends TestCase {
         writeConfigFile("ERR99999", contents);
     }
 
-    @AfterClass
+    @AfterAll
     public void tearDown() throws Exception {
         File appStatus = new File(homeDir + "/app-status.yml");
-        appStatus.delete();
+        if (appStatus.exists()) {
+            appStatus.delete();
+        }
     }
 
     @Test
-    public void testAppStatus() {
+    public void test1AppStatus() {
         config.clear();
         // test default element without merging with app-status
         Status status0 = new Status("ERR10053", "url");
-        Assert.assertEquals(401, status0.getStatusCode());
+        Assertions.assertEquals(401, status0.getStatusCode());
         Server.mergeStatusConfig();
         Status status = new Status("ERR99999");
         System.out.println("****************************************************************" + status.toString());
-        Assert.assertEquals(404, status.getStatusCode());
+        Assertions.assertEquals(404, status.getStatusCode());
         // test default element after merging
         Status status1 = new Status("ERR10053", "url");
-        Assert.assertEquals(401, status1.getStatusCode());
+        Assertions.assertEquals(401, status1.getStatusCode());
     }
 
-    public void testDuplicateStatus() {
+    @Test
+    public void test2DuplicateStatus() {
         config.clear();
         try {
             Server.mergeStatusConfig();
             // second try to make sure duplication status appear
             Server.mergeStatusConfig();
-            fail();
+            fail("Expected RuntimeException was not thrown");
         } catch (RuntimeException expected) {
             // pass
         }
     }
 
-    public void testWithoutAppStatus() {
+    @Test
+    public void test3WithoutAppStatus() {
         config.clear();
         File appStatus = new File(homeDir + "/app-status.yml");
-        appStatus.delete();
+        if (appStatus.exists()) {
+            appStatus.delete();
+        }
         // test default element without merging with app-status
         Status status0 = new Status("ERR10053", "url");
-        Assert.assertEquals(401, status0.getStatusCode());
+        Assertions.assertEquals(401, status0.getStatusCode());
         Server.mergeStatusConfig();
         // test default element after merging
         Status status1 = new Status("ERR10053", "url");
-        Assert.assertEquals(401, status1.getStatusCode());
+        Assertions.assertEquals(401, status1.getStatusCode());
     }
 
-    public void testEmptyAppStatus() {
+    @Test
+    public void test4EmptyAppStatus() throws IOException {
         config.clear();
         File appStatus = new File(homeDir + "/app-status.yml");
-        appStatus.delete();
-        new File(homeDir + "/app-status.yml");
+        if (appStatus.exists()) {
+            appStatus.delete();
+        }
+        appStatus.createNewFile();
         // test default element without merging with app-status
         Status status0 = new Status("ERR10053", "url");
-        Assert.assertEquals(401, status0.getStatusCode());
+        Assertions.assertEquals(401, status0.getStatusCode());
         Server.mergeStatusConfig();
         // test default element after merging
         Status status1 = new Status("ERR10053", "url");
-        Assert.assertEquals(401, status1.getStatusCode());
+        Assertions.assertEquals(401, status1.getStatusCode());
     }
 
     private void setExternalizedConfigDir(String externalizedDir) throws Exception {
